Explicitly define face version
This patch alters the call to `Face.define` to explicitly use "0.0.1" as the
version string. This is required as the code to sort faces by version is
bugged and can't deal with `:current` showing up. This should be safe as the
version "0.0.1" is already defined by the Puppet core and the upload face
does actually call any other faces, so we should avoid the pitfall of
PE-22853 where multiple face versions can cause silent failures.