Add support for PE to acceptance tests
This patch extends the acceptance tests, rake tasks, and TravisCI suite to
install and test the module against PE 2017.3.2 running on CentOS 7. Groundwork
has been laid in the Rakefile for further parameterization of the test matrix.
A notable result of this is that an explicit host configuration file is no
longer used, instead a beaker-hostgenerator string is created and passed to the
--hosts option of the Beaker run.