Update logic for initializing submodules
This commit updates the logic used by Rake tasks for initializing the
`checkouts/puppetserver` submodule. Previously, the task checked for
the existence of the `checkouts/puppetserver` directory. The updated
task check for the existence of the `checkouts/puppetserver/project.clj`
project file as a fresh git clone can leave the directory empty.