Source Puppet Server version from the environment
This commit updates the Rakefile and project.clj to consult a
PUPPETSERVER_VERSION environment variable for which version of
Puppet Server to use. The tasks provided by the Rakefile will also
ensure that the proper version of Puppet Server is checked out and
built into a JAR before items such as tests are run.