Fix dependency for PUPPETSERVER_JAR in Rakefile
This commit updates the dependency for the `PUPPETSERVER_JAR` to be the
`puppetserver:update` task, which ensures the project directory exists
and then ensures the proper commit is checked out. Prior to this change,
the JAR could be built against an incorrect version of the project as
it only require the directory to exist.