Skip mounting services on incompatible versions
This commit adds guard logic to the Clojure FactsUploadService that ensures
routes and status info are not mounted if the plug-in is loaded by an
incompatible Puppet Server version. Compatibility is currently defined by
the X.Y components of the version number and restricted to Puppet Server
5.1 and 5.2.

A test against Puppet Server 5.3 is added to ensure the noop behavior is
triggered.

Ref. issue #2.