Fix Code Manager compatibility
This commit updates the facts_upload::server class to swap the use of a
resource collector for a puppet_enterprise::trapperkeeper::bootstrap_cfg
resource. This construction is much more straight forward and avoids the
effect of PUP-8008, which causes collection of custom resource types to
fail if Code Manager is in use. The acceptance tests are also updated to
run `puppet generate types` as part of module installation to simulate
the effect of Code Manager environment isolation.

Fixes Sharpie/puppet-facts_upload#4.