Clean up test for service compatibility
This commit simplifies the test for service compatibility by replacing the
`puppet apply` change and log assertions with a check for the presence of
the facts-upload-service in /status/v1/services API output. Previously, this
test could fail if the acceptance suite was run multiple times and the
`puppet apply` call had nothing to enforce.