Add partial support for 2016.4 to facts_upload::server
This commit adds support for 2016.4.13 to the facts_upload::server class. The
configuration used to accomplish this won't work for any other PE version older
than 2017.3 due to packaging changes. However 2016.4.13 and newer received
backported changes that allow the Java CLASSPATH of pe-puppetserver to be
extended via a cli-defaults.sh file.