Update PE compatibility in facts_upload::server
In order to enable the facts_upload plugin on PE versions older than 2017.3,
the facts-upload.jar file has to be added to the Java classpath used by
pe-puppetserver. 2016.4.10 had some package changes backported that make this
easy to do with a defaults file. This commit excludes support for 2016.5,
2017.1, 2017.2, and versions of 2016.4 older than 2016.4.10 as the required
packaging change is not present.